# docs/openapi.yaml
openapi: 3.1.0
...
info:
  title: TFB — Landing & Back Office API
  version: 1.0.0
  description: |
    Deux surfaces dans une même application Next.js.

    **API publique** — ce que consomme la landing : le contenu résolu dans une
    langue, les locales actives, le formulaire de contact, le paiement.

    **API admin** (`/api/admin/*`) — protégée par `withAdmin`, gérée par un cookie
    de session signé. Toute route admin non authentifiée répond `401`, avant tout
    accès à la base.

    Deux règles traversent le modèle :

    - **Aucune copy en dur.** Toute chaîne affichée vient de `tfb_translations`.
      Le français est la locale de repli : une traduction absente retombe sur FR,
      et vider une cellule **supprime** la ligne plutôt que d'y stocker `''`.
    - **Aucun média en base.** Les fichiers vont dans `STORAGE_PATH` ; la base ne
      garde que le chemin relatif (`/storage/screenshots/x.png`), servi par
      `GET /api/storage/…`.
  contact:
    name: The Franchise Buddy
servers:
  - url: http://localhost:3000
    description: Développement
  - url: https://{host}
    description: Production
    variables:
      host:
        default: www.franchisebuddy.eu
...
security: []
paths:
...
  /api/admin/packs/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Admin — Packs & licences]
      summary: Un pack, avec le catalogue de modules pour le composer
      operationId: getPack
      x-tfb-status: implemented
      x-tfb-source: src/app/api/admin/packs/[id]/route.ts
      security: [{ adminSession: [] }]
      responses:
        '200':
          description: Le pack et le catalogue.
          content:
            application/json:
              schema:
                type: object
                properties:
                  pack:
                    allOf:
                      - $ref: '#/components/schemas/Pack'
                      - type: object
                        properties:
                          prices:
                            type: array
                            items: { $ref: '#/components/schemas/PackPrice' }
                          moduleIds:
                            type: array
                            items: { type: integer }
                  catalogue:
                    type: array
                    items: { $ref: '#/components/schemas/ModuleLabel' }
                  licenseCount: { type: integer }
                  fixture: { type: boolean }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Admin — Packs & licences]
      summary: Modifie un pack
      description: >-
        `key` est absente du corps : c'est ce sur quoi le code de package d'une
        licence s'aligne, donc la renommer détacherait silencieusement chaque
        licence de ses modules.


        Activer un pack exige un produit Stripe et au moins un prix — sauf pour un
        pack `free`, qui ne facture rien. Sinon la landing afficherait une offre
        que personne ne peut souscrire.
      operationId: updatePack
      x-tfb-status: implemented
      x-tfb-source: src/app/api/admin/packs/[id]/route.ts
      security: [{ adminSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                stripeProductId: { type: [string, 'null'] }
                kind: { $ref: '#/components/schemas/PackKind' }
                icon: { type: [string, 'null'] }
                isActive: { type: boolean }
                isFeatured: { type: boolean }
                sortOrder: { type: integer }
      responses:
        '200':
          description: Pack modifié.
          content:
            application/json:
              schema:
                type: object
                properties:
                  pack: { $ref: '#/components/schemas/Pack' }
        '400':
          description: Produit déjà rattaché, ou activation sans produit ni prix.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Admin — Packs & licences]
      summary: Supprime un pack
      description: >-
        Refusé tant que le service de facturation rapporte des licences sur cette
        clé : chaque magasin concerné perdrait ses modules sans qu'il reste rien
        pour l'expliquer.
      operationId: deletePack
      x-tfb-status: implemented
      x-tfb-source: src/app/api/admin/packs/[id]/route.ts
      security: [{ adminSession: [] }]
      responses:
        '200':
          $ref: '#/components/responses/Ok'
        '400':
          description: Des licences utilisent ce pack.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/admin/licenses/{licenseId}/overrides:
    parameters:
      - $ref: '#/components/parameters/LicenseId'
    get:
      tags: [Admin — Packs & licences]
      summary: Les dérogations d'une licence, en cours et révoquées
      operationId: listOverrides
      x-tfb-status: implemented
      x-tfb-source: src/app/api/admin/licenses/[licenseId]/overrides/route.ts
      security: [{ adminSession: [] }]
      responses:
        '200':
          description: Les dérogations, de la plus récente à la plus ancienne.
          content:
            application/json:
              schema:
                type: object
                properties:
                  overrides:
                    type: array
                    items: { $ref: '#/components/schemas/LicenseOverride' }
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Admin — Packs & licences]
      summary: Pose une dérogation
      description: >-
        Ouvre ou ferme un module hors pack, pour cette licence. C'est le seul
        endroit de la chaîne qu'une décision humaine peut surcharger, et le
        webhook Stripe ne l'écrase pas.


        `reason` est obligatoire : une dérogation sans motif est intraçable six
        mois plus tard.


        Une dérogation existante sur le même module est révoquée et remplacée —
        l'ancienne ligne reste, avec sa date et son auteur.
      operationId: createOverride
      x-tfb-status: implemented
      x-tfb-source: src/app/api/admin/licenses/[licenseId]/overrides/route.ts
      security: [{ adminSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [moduleId, grant, reason]
              properties:
                moduleId: { type: integer, minimum: 1 }
                grant:
                  type: boolean
                  description: '`true` ouvre le module malgré le pack, `false` le ferme malgré le pack.'
                reason: { type: string }
                expiresAt:
                  type: [string, 'null']
                  format: date-time
                  description: Doit être dans le futur. Absente, la dérogation tient jusqu'à sa révocation.
      responses:
        '201':
          description: Dérogation posée.
          content:
            application/json:
              schema:
                type: object
                properties:
                  override: { $ref: '#/components/schemas/LicenseOverride' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Licence ou module introuvable.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  securitySchemes:
    adminSession:
      type: apiKey
      in: cookie
      name: tfb_admin_session
      description: |
        JWT signé (HS256), posé par `POST /api/admin/auth/login`, `httpOnly`,
        `sameSite: lax`, `secure` en production, valable 8 heures.

        Toute route `/api/admin/*` est enveloppée par `withAdmin` : sans cookie
        valide, la réponse est `401` **avant** tout accès à la base.
  parameters:
    LicenseId:
      name: licenseId
      in: path
      required: true
      description: L'identifiant de licence du service de facturation, tel quel.
      schema: { type: string }
    Id:
      name: id
      in: path
      required: true
      description: Entier positif. Toute autre valeur répond `400`.
      schema: { type: integer, minimum: 1 }
  responses:
    Ok:
      description: Opération effectuée.
      content:
        application/json:
          schema:
            type: object
            properties:
              ok: { type: boolean, const: true }
    BadRequest:
      description: Requête invalide.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Authentification requise.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Ressource introuvable.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Message en français, destiné à être affiché tel quel.
...
    PackKind:
      type: string
      enum: [subscription, one_time, addon, free]
      default: subscription
      description: Comment le pack est facturé — pas ce qu'il ouvre.
    Pack:
      type: object
      properties:
        id: { type: integer }
        key:
          type: string
          description: Égale au `code` du package côté service de facturation. Immuable.
        stripeProductId: { type: [string, 'null'] }
        kind: { $ref: '#/components/schemas/PackKind' }
        icon: { type: [string, 'null'] }
        isActive: { type: boolean }
        isFeatured: { type: boolean }
        sortOrder: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    PackPrice:
      type: object
      properties:
        id: { type: integer }
        packId: { type: integer }
        stripePriceId: { type: string }
        amount: { type: [integer, 'null'], description: En centimes. `null` = sur devis. }
        currency: { type: string }
        interval: { type: string, enum: [month, year, one_time] }
        isDefault: { type: boolean }
    ModuleLabel:
      type: object
      properties:
        id: { type: integer }
        key: { type: string }
        name: { type: string, description: Libellé français, ou la clé tant qu'aucune traduction n'existe. }
        icon: { type: [string, 'null'] }
        group: { type: [string, 'null'] }
        isActive: { type: boolean }
...
    LicenseOverride:
      type: object
      properties:
        id: { type: integer }
        licenseId:
          type: string
          description: L'identifiant du service de facturation. Aucune clé étrangère ne traverse cette frontière, volontairement.
        moduleId: { type: integer }
        grant: { type: boolean }
        reason: { type: string }
        expiresAt: { type: [string, 'null'], format: date-time }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
        revokedAt: { type: [string, 'null'], format: date-time }
        revokedBy: { type: [string, 'null'] }
...